refactor(php,python): split adapter packaging and reporting seams
